requirements.txt
backcall==0.1.0
bcrypt==3.1.7
bip32utils==0.3.post4
cffi==1.14.0
decorator==4.4.2
ecdsa==0.15
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
loguru==0.4.1
mnemonic==0.19
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
ptyprocess==0.6.0
pycparser==2.19
pycryptodome==3.9.7
Pygments==2.5.2
six==1.14.0
traitlets==4.3.3
wcwidth==0.1.8
